Split tooltip into base tooltip to allow for custom content
