+++ libs/@hashintel/ds-components/src/components/Tooltip/tooltip.recipe.ts
@@ -1,19 +1,4 @@
-import { css, cva } from "@hashintel/ds-helpers/css";
-
-export const triggerStyles = css({
-  lineHeight: "[0]",
-
-  "&:focus-visible": {
-    outline: "[2px solid]",
-    outlineColor: "neutral.s30",
-    outlineOffset: "[2px]",
-    borderRadius: "md",
-  },
-});
-
-export const positionerStyles = css({
-  zIndex: "tooltip !important",
-});
+import { cva } from "@hashintel/ds-helpers/css";
 
 export const contentStyles = cva({
   base: {
